resell-bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
python-telegram-bot==20.7
requests-oauthlib==1.3.1
rapidfuzz==3.6.1
lxml==5.1.0
